Handle possible exception of MIDI not available

+++ src/main/java/com/example/pianino/ui/SettingsController.java
@@ -64,7 +64,7 @@ public class SettingsController {
         velocitySlider.setBlockIncrement(1);
 
         velocitySlider.valueProperty().addListener((obs, oldValue, newValue) -> {
-            piano.setVelocity(newValue.byteValue());
+            piano.setVelocity(newValue.intValue());
         });
 
         root.getChildren().addAll(
